requirements.txt
discord.py

python-dotenv

tinydb

validator-collection

pytz
